Fix for Issue #3
This fixes issue #3 : When running the removeAllUsersFromDB proc, it will throw an error if it tries to remove a user that is the owner of a schema.
This was fixed by resetting any "non-dbo" schema owner back to "dbo".  Since all users are going to be dropped anyway, this is necessary to 1) prevent the error, and 2) ensure that the schema and any database objects (tables, procs, etc...) possibly assigned to the schema remain intact.